fix constant divisions (#89)
Constant divisions were optimised using multiplications, but the result
was sometimes off in the least significant digit. Reverting to a slower
but more precise algo.

Also fix a bug in divrem64, which checked the wrong input for lo == 0.
This was previously benign, since the input and output arguments were
always the same variable.